+++ lib/cli.ex
@@ -1,3 +1,0 @@
-defmodule LastfmArchive.Cli do
-  @moduledoc false
-end