+++ test/lastfm_archive/behaviour/archive_test.exs
@@ -1,5 +1,6 @@
-defmodule Lastfm.ArchiveTest do
+defmodule LastfmArchive.Behaviour.ArchiveTest do
   use ExUnit.Case, async: true
+  alias LastfmArchive.Behaviour.Archive
 
   test "new/1 returns expected archive metadata " do
     assert %{
@@ -10,6 +11,6 @@ defmodule Lastfm.ArchiveTest do
              identifier: "a_lastfm_user",
              source: "http://ws.audioscrobbler.com/2.0",
              title: "Lastfm archive of a_lastfm_user"
-           } = Lastfm.Archive.new("a_lastfm_user")
+           } = Archive.new("a_lastfm_user")
   end
 end