+++ test/support/lastfm_client_stub.ex
@@ -1,7 +1,7 @@
-defmodule Lastfm.ClientStub do
+defmodule LastfmArchive.LastfmClientStub do
   @moduledoc false
 
-  @behaviour Lastfm.Client
+  @behaviour LastfmArchive.Behaviour.LastfmClient
 
   @registered_time DateTime.from_iso8601("2021-04-01T18:50:07Z") |> elem(1) |> DateTime.to_unix()
   @latest_scrobble_time DateTime.from_iso8601("2021-04-03T18:50:07Z") |> elem(1) |> DateTime.to_unix()