+++ test/support/mocks.ex
@@ -1,5 +1,7 @@
-Mox.defmock(Lastfm.ClientMock, for: Lastfm.Client)
-Mox.defmock(Lastfm.FileArchiveMock, for: Lastfm.Archive)
-Mox.defmock(Lastfm.FileIOMock, for: Lastfm.FileIO)
-Mox.defmock(Lastfm.PathIOMock, for: Lastfm.PathIO)
-Mox.defmock(LastfmArchive.CacheMock, for: LastfmArchive.Cache)
+alias LastfmArchive.Behaviour
+
+Hammox.defmock(LastfmArchive.LastfmClientMock, for: Behaviour.LastfmClient)
+Hammox.defmock(LastfmArchive.FileArchiveMock, for: Behaviour.Archive)
+Hammox.defmock(LastfmArchive.FileIOMock, for: Behaviour.FileIO)
+Hammox.defmock(LastfmArchive.PathIOMock, for: Behaviour.PathIO)
+Hammox.defmock(LastfmArchive.CacheMock, for: LastfmArchive.Cache)